// src/types/types.ts
export type MenuType = {
    id: string;
    title: string;
    desc?: string;
    color: string;
    img?: string;
    slug: string
}[];

export type ProductType = {
    id: string;
    title: string;
    desc: string;
    img?: string;
    price: number;
    options?: { title: string; additionalPrice: number }[];
}

export type OrderType = {
    id: string;
    userEmail: string;
    price: number;
    products: CartItemType[];
    status: string;
    createdAt: Date;
    intent_id?: String
}

export type CartItemType = {
    id: string;
    title: string;
    img?: string;
    price: number;
    optionTitle?: string;
    quantity: number
}

export type CartType = {
    products:CartItemType [];
    totalItems: number;
    totalPrice: number
}

export type ActionTypes = {
    addToCart: (item: CartItemType) => void;
    removeFromCart: (item: CartItemType) => void;
}